refactor: extract SharedComponents to separate file
- Move SharedComponents and SharedComponent from factory.go to sharedcomponent.go
- Add attribution comment pointing to OpenTelemetry collector-contrib source
- Improve separation of concerns between factory creation and lifecycle management
- All tests pass including integration tests and Docker build
